+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 ## 53.0 - 2023-xx-xx
 ### Added
 - See the characters with the most progress for a given achievement in its tooltip
+- Additional options and data management
 
 ### Changed
 - Custom objectives are now handled in a different way to allow more flexibility in the future
 
 ### Fixed
-- Achievement objectives in the tooltip for completed achievements should once again listen to the settings 
+- Achievement objectives in the tooltip for completed achievements should once again listen to the settings
+
+### Wrath Classic
+- Fixed an error with the Keybindings localization
 
 ## 52.4 - 2023-02-23
 ### Wrath Classic